fixup: modules: separate library into modules
Separates the library into modules for better readability and
maintainability.
